[Core/LuaEngine] Fix LuaEngine windows build
