+++ .github/workflows/windows-build.yml
@@ -70,7 +70,7 @@ jobs:
         cd $GITHUB_WORKSPACE
         
         mkdir -p build && cd build
-        cmake .. -DTOOLS=ON -DELUNA=0 -A x64
+        cmake .. -DTOOLS=ON -DELUNA=1 -A x64
         cmake --build . --config RelWithDebInfo --parallel 4
       #git bash shell
       shell: bash
